fix: Report 점포수 Api 쿼리를 Desc 으로 변경, response Dto 배열을 long 타입으로 변경

+++ backend/bizscanner/src/main/java/store/bizscanner/dto/response/store/QuarterlyStoreResponse.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 @Getter
 public class QuarterlyStoreResponse {
-    private final Integer[] quarterlyStore;
+    private final Long[] quarterlyStore;
 
     public QuarterlyStoreResponse(List<TotalStoreMapping> quarterlyStoreList) {
-        quarterlyStore = quarterlyStoreList.stream().map(TotalStoreMapping::getStoreCount).toArray(Integer[]::new);
+        quarterlyStore = quarterlyStoreList.stream().map(TotalStoreMapping::getStoreCount).toArray(Long[]::new);
     }
 }
